fix: lint && clean code
