+++ packages/components/src/Token/index.tsx
@@ -93,7 +93,7 @@ const TokenIcon = ({
         alt={network.logoURI}
         borderRadius="full"
         borderWidth="2px"
-        borderColor="#000"
+        borderColor="surface-subdued"
         position="absolute"
         top="-4px"
         right="-4px"
