+++ packages/kit/src/views/PushNotification/PriceAlertListModal.tsx
@@ -10,16 +10,17 @@ import { PriceAlertItem } from '@onekeyhq/engine/src/managers/notification';
 import debugLogger from '@onekeyhq/shared/src/logger/debugLogger';
 
 import backgroundApiProxy from '../../background/instance/backgroundApiProxy';
+import { useSettings } from '../../hooks/redux';
+import { setPushNotificationConfig } from '../../store/reducers/settings';
 import {
   ManageTokenRoutes,
   ManageTokenRoutesParams,
 } from '../ManageTokens/types';
 
-import { Item, ListEmptyComponent } from './PriceAlertListStack';
+import { ListEmptyComponent } from './PriceAlertListStack';
+import PriceItem from './PriceItem';
 
 import type { NativeStackNavigationProp } from '@react-navigation/native-stack';
-import { useSettings } from '../../hooks/redux';
-import { setPushNotificationConfig } from '../../store/reducers/settings';
 
 type NavigationProps = NativeStackNavigationProp<
   ManageTokenRoutesParams,
@@ -39,11 +40,11 @@ export const PriceAlertListModal: FC = () => {
   const [loading, setLoading] = useState(false);
   const [data, setData] = useState<PriceAlertItem[]>([]);
   const route = useRoute<RouteProps>();
-  const {pushNotification} = useSettings();
+  const { pushNotification } = useSettings();
   const { token } = route.params;
   const navigation = useNavigation<NavigationProps>();
 
-  const {engine, dispatch} = backgroundApiProxy;
+  const { engine, dispatch } = backgroundApiProxy;
 
   const disabled = useMemo(() => data.length >= MAX_ALERT, [data.length]);
 
@@ -68,17 +69,17 @@ export const PriceAlertListModal: FC = () => {
         setLoading(false);
       }, 100);
     },
-    [token],
+    [token, engine],
   );
 
   const onPrimaryActionPress = useCallback(() => {
-    if(!pushNotification?.pushEnable){
-      dispatch(setPushNotificationConfig({pushEnable: true}));
+    if (!pushNotification?.pushEnable) {
+      dispatch(setPushNotificationConfig({ pushEnable: true }));
     }
     navigation.navigate(ManageTokenRoutes.PriceAlertAdd, {
       token,
     });
-  }, [pushNotification?.pushEnable])
+  }, [pushNotification?.pushEnable, dispatch, navigation, token]);
 
   useFocusEffect(
     useCallback(() => {
@@ -105,7 +106,7 @@ export const PriceAlertListModal: FC = () => {
         borderColor="border-subdued"
       >
         {data.map((item) => (
-          <Item
+          <PriceItem
             alert={item}
             token={token}
             divider={false}
