+++ packages/kit/src/views/PushNotification/PriceAlertListStack.tsx
@@ -1,37 +1,26 @@
-import React, {
-  FC,
-  useCallback,
-  useLayoutEffect,
-  useMemo,
-  useState,
-} from 'react';
+import React, { FC, useCallback, useLayoutEffect, useMemo } from 'react';
 
-import { groupBy, pick } from 'lodash';
-import { FormattedNumber, useIntl } from 'react-intl';
+import { groupBy } from 'lodash';
+import { useIntl } from 'react-intl';
 
 import {
   Box,
   Center,
   Empty,
-  IconButton,
   ScrollView,
   Spinner,
   Text,
-  Token,
   useTheme,
 } from '@onekeyhq/components';
 import { PriceAlertItem } from '@onekeyhq/engine/src/managers/notification';
-import { Token as TokenType } from '@onekeyhq/engine/src/types/token';
 import imageUrl from '@onekeyhq/kit/assets/alert.png';
 import { useNavigation } from '@onekeyhq/kit/src/hooks';
-import debugLogger from '@onekeyhq/shared/src/logger/debugLogger';
 
-import backgroundApiProxy from '../../background/instance/backgroundApiProxy';
 import { useSingleToken } from '../../hooks/useTokens';
 import { HomeRoutes, HomeRoutesParams } from '../../routes/types';
-import { getSuggestedDecimals } from '../../utils/priceUtils';
 
 import { usePriceAlertlist } from './hooks';
+import PriceItem from './PriceItem';
 
 import type { NativeStackNavigationProp } from '@react-navigation/native-stack';
 
@@ -76,97 +65,6 @@ export const ListEmptyComponent: FC<ListEmptyComponentProps> = ({
     />
   );
 };
-export function FormatCurrencyNumber({
-  decimals,
-  value,
-  onlyNumber,
-  currency,
-}: {
-  value: number;
-  decimals?: number;
-  onlyNumber?: boolean;
-  currency: string;
-}) {
-  if (typeof value !== 'number') {
-    return null;
-  }
-  const maxDecimals = decimals ?? getSuggestedDecimals(value);
-
-  return (
-    <FormattedNumber
-      value={value}
-      currencyDisplay="narrowSymbol"
-      // eslint-disable-next-line react/style-prop-object
-      style={onlyNumber ? 'decimal' : 'currency'}
-      minimumFractionDigits={2}
-      maximumFractionDigits={maxDecimals}
-      currency={currency}
-    />
-  );
-}
-
-export const Item: FC<{
-  divider: boolean;
-  onRemove: (price: string) => void;
-  alert: PriceAlertItem;
-  token: TokenType;
-}> = ({ divider, token, alert, ...props }) => {
-  const { price, currency } = alert;
-  const [loading, setLoading] = useState(false);
-
-  const onRemove = useCallback(async () => {
-    setLoading(true);
-    try {
-      await backgroundApiProxy.engine.removePriceAlertConfig(
-        pick(alert, 'impl', 'chainId', 'price', 'currency', 'address'),
-      );
-    } catch (error) {
-      debugLogger.common.error(
-        `remove PriceAlert`,
-        error instanceof Error ? error?.message : error,
-      );
-    }
-    setTimeout(() => {
-      setLoading(false);
-      props?.onRemove(alert.price);
-    }, 200);
-  }, [props, alert]);
-
-  return (
-    <Box
-      display="flex"
-      flexDirection="row"
-      justifyContent="space-between"
-      alignItems="center"
-      py={4}
-      px={{ base: 4, md: 6 }}
-      borderBottomWidth={divider ? '1 ' : undefined}
-      borderBottomColor="divider"
-      borderBottomRadius={divider ? undefined : '12px'}
-    >
-      <Box flex="1" flexDirection="row" alignItems="center">
-        <Token src={token.logoURI} size={8} networkId={token.networkId} />
-        <Text typography="Body1Strong" numberOfLines={1} ml={3}>
-          {`1 ${token.symbol} = `}
-          <FormatCurrencyNumber value={+(price || 0)} currency={currency} />
-        </Text>
-      </Box>
-      <Box w="8" h="8" alignItems="center" justifyContent="center">
-        {loading ? (
-          <Spinner size="sm" />
-        ) : (
-          <IconButton
-            name="TrashSolid"
-            type="plain"
-            circle
-            onPress={onRemove}
-          />
-        )}
-      </Box>
-    </Box>
-  );
-};
-
 const Section = ({
   alerts,
   onRemove,
@@ -205,7 +103,7 @@ const Section = ({
         borderColor="border-subdued"
       >
         {alerts.map((a, index) => (
-          <Item
+          <PriceItem
             alert={a}
             token={token}
             key={`${a.price}-${a.currency}`}
